+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/openfga/go-sdk/compare/v0.7.4...HEAD)
 
+- fix: process credentials when custom HTTPClient is provided (#276)
+
 ## v0.7.4
 
 ### [0.7.4](https://github.com/openfga/go-sdk/compare/v0.7.3...v0.7.4) (2026-01-21)